Refactor routes so registration has its own route
