+++ app/app.js
@@ -6,6 +6,7 @@ var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
 
 var index = require('./routes/index');
+var register = require('./routes/register');
 var users = require('./routes/users');
 
 var app = express();
@@ -27,6 +28,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', index);
+app.use('/register', register);
 app.use('/users', users);
 
 // catch 404 and forward to error handler
